% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.create_individual_plots}
\alias{.create_individual_plots}
\title{Create individual plots for specified metabolites - FIXED AXES}
\usage{
.create_individual_plots(
  data_PP,
  data_DR,
  data_FC,
  data_CA,
  plot_iden_met,
  group1,
  group2,
  group_name,
  groups,
  non_qc_indices,
  direction,
  confidence_level
)
}
\description{
Create individual plots for specified metabolites - FIXED AXES
}
\keyword{internal}
