% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.create_roc_plot}
\alias{.create_roc_plot}
\title{Create ROC plot for top features - FIXED AXES}
\usage{
.create_roc_plot(
  feature_data,
  auroc_df,
  group1,
  group2,
  group_name,
  direction,
  top_n
)
}
\description{
Create ROC plot for top features - FIXED AXES
}
\keyword{internal}
