% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.extract_and_merge_data}
\alias{.extract_and_merge_data}
\title{Extract and merge VIP, fold change, and comparative analysis data}
\usage{
.extract_and_merge_data(data_DR, data_FC, data_CA, group_name)
}
\description{
Extract and merge VIP, fold change, and comparative analysis data
}
\keyword{internal}
