% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.filter_features}
\alias{.filter_features}
\title{Filter features based on specified criteria}
\usage{
.filter_features(merged_data, VIPmin, fcUP, fcDown, adjpvalue)
}
\description{
Filter features based on specified criteria
}
\keyword{internal}
