% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.get_data_matrix_AUROC}
\alias{.get_data_matrix_AUROC}
\title{Get appropriate data matrix based on auto_merge_replicates setting}
\usage{
.get_data_matrix_AUROC(data_PP)
}
\description{
Get appropriate data matrix based on auto_merge_replicates setting
}
\keyword{internal}
