% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.perform_auroc_analysis}
\alias{.perform_auroc_analysis}
\title{Perform AUROC analysis for filtered features}
\usage{
.perform_auroc_analysis(
  feature_data,
  group1,
  group2,
  direction,
  confidence_level
)
}
\description{
Perform AUROC analysis for filtered features
}
\keyword{internal}
