% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.prepare_feature_data}
\alias{.prepare_feature_data}
\title{Prepare feature data matrix for AUROC analysis}
\usage{
.prepare_feature_data(data_PP, feature_names, groups, non_qc_indices)
}
\description{
Prepare feature data matrix for AUROC analysis
}
\keyword{internal}
