% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{.process_group_data_auroc}
\alias{.process_group_data_auroc}
\title{Process group data and handle QC samples}
\usage{
.process_group_data_auroc(data_PP, arrangeLevels, min_group_size)
}
\description{
Process group data and handle QC samples
}
\keyword{internal}
