% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{extract_auroc_results}
\alias{extract_auroc_results}
\title{Extract AUROC Results}
\usage{
extract_auroc_results(
  auroc_results,
  comparison = NULL,
  min_auc = 0,
  max_auc = 1,
  top_n = NULL,
  include_merged_data = FALSE,
  sort_by = "auc_desc"
)
}
\description{
Extract AUROC Results
}
