% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{extract_file_info}
\alias{extract_file_info}
\title{Extract File Information}
\usage{
extract_file_info(file_location, raw_data)
}
\description{
Extract File Information
}
\keyword{internal}
