% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_PCA.R
\name{get_ScreeData}
\alias{get_ScreeData}
\title{Extract Scree Plot Data}
\usage{
get_ScreeData(multi_plots_obj, what = "variance", max_pc = NULL)
}
\arguments{
\item{multi_plots_obj}{Object returned from \code{perform_PCA()}.}

\item{what}{Character. \code{"variance"} or \code{"eigenvalue"}.}

\item{max_pc}{Integer. Maximum number of PCs. If \code{NULL}, includes all.}
}
\value{
Data frame with PC-level summary statistics.
}
\description{
Extract Scree Plot Data
}
