% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{load_and_validate_file}
\alias{load_and_validate_file}
\title{Load and Validate File}
\usage{
load_and_validate_file(
  file_location,
  sheet_name,
  skip_rows,
  separator,
  verbose
)
}
\description{
Load and Validate File
}
\keyword{internal}
