% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{merge_duplicate_samples}
\alias{merge_duplicate_samples}
\title{Merge Duplicate Samples}
\usage{
merge_duplicate_samples(
  processed_data,
  method,
  metadata_cols_processed,
  feature_cols_processed,
  verbose
)
}
\description{
Merge Duplicate Samples
}
\keyword{internal}
