% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performDataQualityCheck.R
\name{performDataQualityCheck}
\alias{performDataQualityCheck}
\title{Import an Excel (.xlsx or .csv, or .tsv) File or a Text (.txt) File}
\usage{
performDataQualityCheck(
  file_location = NULL,
  sheet_name = NULL,
  skip_rows = 0,
  separator = ","
)
}
\arguments{
\item{file_location}{String. A string of the location of the data in the device. Must have the following characteristics:
\itemize{
\item 1st row name is "Sample": The sample names, ideally shortened versions to make it easier to view in visualizations; without spaces.
\item 2nd row name is "SubjectID": Must be numeric to be sorted correctly. Can contain non-unique identifiers of "Sample" names, usually used when "Replicate" is present. Also referred to as the individuals in the data.
\item 3rd row name is "Replicate": The replicate IDs of Samples. Ideally non-unique identifiers.
\item 4th row name is "Group": Contains QC and the groups (with/out spaces).
\item 5th row name is "Batch": The batch numbers.
\item 6th row name is "Injection": The injection sequence.
\item 7th row name is "DilutionMarker": Concentration markers to be used during data normalization to correct for variations in e.g., urine, concentration between samples. Can be osmolality values, specific gravity, etc.
\item 8th row name is "Response": Numeric. Ideally used when the the response variable (dependent variable) to be used is not "Group".
\item Others requirement/s are:
\itemize{
\item The 1st 8 rows above must be present.
\item After the 8th row (Response), are the features possibly in a mass-to-charge ratio (m/z) and retention time format. Example \verb{199.101@0.111} assuming you rounded off the decimal places to 3 decimal places (suggested for a cleaner output in visualizations).
\item If the DilutionMarker values, SubjectID, Replicate, and Response are not present, leave the rows blank except for the words in the 1st column (e.g., "DilutionMarker' ).
\item Missing values must be left blank or encoded as 0. Blanks and 0s are identified as NAs or missing values.
}
}}

\item{sheet_name}{String. The name of the worksheet where the data is located. Ignored when data type is ".csv".}

\item{skip_rows}{Numeric. The number of rows in the data to skip reading.}

\item{separator}{String. The separator in the ".csv" or ".txt" file. If this fails, it's likely that the file is supposed to be delimited, which should be saved as ".csv", ".tsv", or ".txt".
\itemize{
\item ",": A comma. From a comma-separated file (csv).
\item "/t": A tab From a tab-separated file (.txt or .tsv).
\item Also accepts other delimiters, as long as it is consistent in your data.
}
Defaults to ",". Ignored when the file is ".xlsx".}
}
\value{
A data frame ready to be used in the \code{performPreprocessingPeakData} function in the \code{MetaboStatR} package.
}
\description{
This function imports Excel or text files and performs quality controls prior to usage in the
succeeding functions. This is ideally used to check data requirements such as uniqueness of
inputs in the "Sample" row, uniqueness of Features/Metabolites, data types, and the
overall characteristics of the data.
}
\examples{
\dontrun{
# Using the default parameters
# The location typically looks like
# "C:/Users/Documents/Metabolomics Data Analysis/file_name_of_data.xlsx"
performDataQualityCheck(
 file_location = the_complete_location_of_your_data
)
}

}
