% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_FoldChange.R
\name{perform_FoldChange}
\alias{perform_FoldChange}
\title{Perform Fold Change Analysis on a Preprocessed Data}
\usage{
perform_FoldChange(data, arrangeLevels = NULL, sortFC = TRUE)
}
\arguments{
\item{data}{List. This list must be a result from the \code{perform_PreprocessingPeakData} function. It can also be from a data scaling technique.}

\item{arrangeLevels}{Vector. Determines how the groups will be arranged. The format could be "c('group1', 'group2', ...)". Defaults to \code{NULL} which sorts the groups in alphabetical order.}

\item{sortFC}{Boolean. If \code{TRUE} (default), sorts the fold changes in descending order.}
}
\value{
A list of data frames.
}
\description{
This function performs fold change analysis on a preprocessed Data.
}
\examples{
\dontrun{
perform_FoldChange(data = results_from_perform_PreprocessingPeakData_function)
}

}
