% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAUROC.R
\name{performAUROC}
\alias{performAUROC}
\title{Perform Area Under the Receiver Operating Characteristic (AUROC)}
\usage{
performAUROC(
  data_PP,
  data_DR,
  data_FC,
  data_CA,
  arrangeLevels = NULL,
  VIPmin = 1,
  fcUP = 2,
  fcDown = 0.5,
  adjpvalue = 0.05,
  direction = "auto2",
  top_n = 5,
  plot_iden_met = NULL
)
}
\arguments{
\item{data_PP}{List. This list must be a result from the \code{performPreprocessingPeakData} function.}

\item{data_DR}{List. This list must be a result from the \code{performDimensionReduction} function, specifically from OPLS-DA.}

\item{data_FC}{List. This list must be a result from the \code{performFoldChange} function.}

\item{data_CA}{List. This list must be a result from the \code{performComparativeAnalysis} function.}

\item{arrangeLevels}{Vector. Determines how the groups will be arranged. The format could be "c('group1', 'group2', ...)". Defaults to \code{NULL} which sorts the groups in alphabetical order. Suggests to be inputted (control, case1, case2) where case2 is worse than case1 e.g., Severe dengue than Non-severe dengue.}

\item{VIPmin}{Numeric. Minimum VIP score to consider.}

\item{fcUP}{Numeric. Minimum fold change to consider feature as up-regulated.}

\item{fcDown}{Numeric. Maximum fold change to consider feature as down-regulated.}

\item{adjpvalue}{Numeric. Defaults to 0.05. Defines the adjusted p-value threshold.}

\item{direction}{Defines the direction in AUROC.
\itemize{
\item "auto": (This is not applied.)
\item ">": median of control < median of case
\item "<": median of control > median of case
\item "auto2": Dynamically change the direction. If \verb{median of control < median of case}, then \code{>} will be applied, ottherwise, \code{<}.
}
Defaults to ">".}

\item{top_n}{Numeric. Defines the features with the highest AUC. Defaults to plotting the top 5.}

\item{plot_iden_met}{Boolean. If \code{TRUE}, plots the identified metabolites, one-by-one. Defaults to \code{NULL}.}
}
\value{
Returns a list of results and plots if requested.
}
\description{
This function performs the AUROC analysis using the data from data processing, dimension-reduction,
fold change analysis, and comparative analysis. Data will be automatically selected from those lists.
}
\examples{
\dontrun{
  data_PP = results_from_performPreprocessingPeakData_function,
  data_DR = results_from_performDimensionReduction_function,
  data_FC = results_from_performFoldChange_function,
  data_CA = results_from_performComparativeAnalysis_function
}
}
