% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performDimensionReduction.R
\name{performDimensionReduction}
\alias{performDimensionReduction}
\title{Function to Perform Dimension Reduction Techniques on the Processed Data}
\usage{
performDimensionReduction(
  data,
  type = "PCA",
  reduceWhat = "QC",
  arrangeLevels = NULL,
  screeWhat = "variance",
  screeType = "both",
  screeLabs = TRUE,
  screeTitle = "",
  screeXLab = "Principal Components",
  screeYLab = "\% Variance Explained",
  scoresEllipse = TRUE,
  scoresTitle = NULL,
  scoresLegend = NULL
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{performPreprocessingPeakData} function.}

\item{type}{String. The type of dimension reduction technique to perform. Choices are below. Future choices will include "PLS-DA" and "sPLS-DA".
\itemize{
\item "PCA": Perform Principal Component Analysis (PCA).
\item "OPLS-DA": Perform Orthogonal Partial Least Squares-Discriminant Analysis (OPLS-DA).
}
Defaults to "PCA".}

\item{reduceWhat}{String. Controls what you want to perform the selected \code{type} on.
\itemize{
\item "QC": Perform the selected \code{type} on QC = Quality Control Samples.
\item "BS": Perform the selected \code{type} on BS = Biological Samples.
}
Defaults to "QC".}

\item{arrangeLevels}{Vector. Determines how the groups will be arranged. The format could be "c('group1', 'group2', ...)". Defaults to \code{NULL} which sorts the groups in alphabetical order.}

\item{screeWhat}{String. Determines what will be plotted in the scree plot (also called elbow plot).
\itemize{
\item "variance": Plot the variance explained.
\item "eigenvalue": Plot the eigen values.
}
Defaults to "variance".}

\item{screeType}{String. Controls how the scree plot is plotted.
\itemize{
\item "bar": Uses bar plots.
\item "line": Uses line plots.
\item "both": Uses both bar and line plots.
}
Defaults to "both".}

\item{screeLabs}{Boolean. If \code{TRUE} (default), adds labels.}

\item{screeTitle}{String. The scree plot title. If \code{NULL} or empty (""), defaults to "Scree plot".}

\item{screeXLab}{String. The scree plot x-axis label. Defaults to "Principal Components".}

\item{screeYLab}{String. The scree plot y-axis label. Defaults to "\% Variance Explained".}

\item{scoresEllipse}{Boolean. If \code{TRUE} (default), adds an ellipse in the scores plot.}

\item{scoresTitle}{String. The scores plot title.}

\item{scoresLegend}{String. The title in the legend section of the scores plot. Defaults to \code{NULL} which means no legend title.}
}
\value{
Returns a range of list of results from plots, data frames, etc.
}
\description{
This function performs dimension reduction techniques on the preprocessed data. These
techniques include Principal Component Analysis (PCA). and
Orthogonal Partial Least Squares-Discriminant Analysis (OPLS-DA). Techniques to be added
are Partial Least Squares-Discriminant Analysis (PLS-DA) and
sparse Partial Least Squares-Discriminant Analysis (sPLS-DA). Parameters will be ignored
if they are not usable in the 'type' indicated.
}
\examples{
\dontrun{
performDimensionReduction(
  data = data_from_performPreprocessingPeakData_function
)
}
}
