% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performExport2Excel.R
\name{performExport2Excel}
\alias{performExport2Excel}
\title{Perform Export Data Frames to Excel}
\usage{
performExport2Excel(
  results,
  folder_name = "Results_Folder",
  file_name = "Results Exports"
)
}
\arguments{
\item{results}{List. A list of results from \code{performPreprocessingPeakData}, \code{performDimensionReduction}, or \code{performComparativeAnalysis}, etc. There must only be one at a time. The list may or may not contain a data frame.}

\item{folder_name}{String. Creates a folder in the current working directory where the Excel file will be exported.}

\item{file_name}{String. The file name of the results.}
}
\value{
An Excel file.
}
\description{
This function exports the data frames from a list to Excel. The data frames are detected automatically. All data frames
in a list are combined in one (1) Excel file.
}
\examples{
\dontrun{
performExport2Excel(
  results     = results_from_performRegression_function,
  folder_name = "Regression Results",
  file_name   = "Regression Results")
}

}
