% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performFoldChange.R
\name{performFoldChange}
\alias{performFoldChange}
\title{Perform Fold Change Analysis on a Preprocessed Data}
\usage{
performFoldChange(data, arrangeLevels = NULL, sortFC = TRUE)
}
\arguments{
\item{data}{List. This list must be a result from the \code{performPreprocessingPeakData} function. It can also be from a data scaling technique.}

\item{arrangeLevels}{Vector. Determines how the groups will be arranged. The format could be "c('group1', 'group2', ...)". Defaults to \code{NULL} which sorts the groups in alphabetical order.}

\item{sortFC}{Boolean. If \code{TRUE} (default), sorts the fold changes in descending order.}
}
\value{
A list of data frames.
}
\description{
This function performs fold change analysis on a preprocessed Data.
}
\examples{
\dontrun{
performFoldChange(data = results_from_performPreprocessingPeakData_function)
}

}
