% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_AUROC.R
\name{plot_auroc_distribution}
\alias{plot_auroc_distribution}
\title{Plot AUROC distribution}
\usage{
plot_auroc_distribution(auroc_results, bins = 20)
}
\arguments{
\item{auroc_results}{List returned by perform_AUROC}

\item{bins}{Integer. Number of histogram bins}
}
\value{
ggplot2 object
}
\description{
Plot AUROC distribution
}
