% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CorrelationHeatmap.R
\name{plot_CorrelationHeatmap}
\alias{plot_CorrelationHeatmap}
\title{Plot Correlational Heatmaps}
\usage{
plot_CorrelationHeatmap(
  data,
  method = c("pearson", "spearman", "kendall")[1],
  plot_top_n = 1000,
  plot_what = c("Samples", "Features")[2],
  plot_type = c("correlation", "hierarchical")[1],
  show_rownames = TRUE,
  show_colnames = TRUE,
  clustering_distance_rows = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski")[1],
  clustering_distance_cols = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski")[1],
  clustering_method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid")[1]
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{perform_PreprocessingPeakData} function. It can also be from a data scaling technique.}

\item{method}{String. Specifies the correlation method to be used.
\itemize{
\item "pearson": Uses Pearson r.
\item "spearman": Uses Spearman rho.
\item "kendall": Uses Kendall tau.
}
Defaults to "pearson".}

\item{plot_top_n}{Numeric. Specifies the top n features/samples to be plotted.}

\item{plot_what}{String. Specifies if Sample or Features are to be plotted.
\itemize{
\item "Samples": Plots the samples.
\item "Features": Plots the features.
}
Defaults to "Features".}

\item{plot_type}{String. Specifies the type of plot.
\itemize{
\item "correlation": Correlational plot.
\item "hierarchical": Hierarchical plot.
}
Defaults to "correlation".}

\item{show_rownames}{Boolean. If \code{TRUE} (default), shows the row names.}

\item{show_colnames}{Boolean. If \code{TRUE} (default), shows the column names.}

\item{clustering_distance_rows}{String. Specifies the clustering distance in the rows. See \code{?dist} for more info.
\itemize{
\item "euclidean": Usual distance between the two vectors.
\item "maximum": Maximum distance between two components of \code{x} and \code{y} (supremum norm)
\item "manhattan": Absolute distance between the two vectors (1 norm aka L_1). (L_1 = L subscript 1)
\item "canberra": Terms with zero numerator and denominator are omitted from the sum and treated as if the values were missing. This is intended for non-negative values (e.g., counts), in which case the denominator can be written in various equivalent ways.
\item "binary": (aka asymmetric binary): The vectors are regarded as binary bits, so non-zero elements are ‘on’ and zero elements are ‘off’. The distance is the proportion of bits in which only one is on amongst those in which at least one is on. This also called “Jaccard” distance in some contexts. Here, two all-zero observations have distance 0, whereas in traditional Jaccard definitions, the distance would be undefined for that case and give NaN numerically.
\item "minkowski": The \code{p} norm, the \code{p-th} root of the sum of the \code{p-th} powers of the differences of the components.
}
Defaults to "euclidean".}

\item{clustering_distance_cols}{String. Specifies the clustering distance in the columns. Same as in \code{clustering_distance_rows}. Defaults to "euclidean".}

\item{clustering_method}{Specifies the clustering method. See \code{?hclust} for more info.
\itemize{
\item "ward.D": Ward's minimum variance method aims at finding compact, spherical clusters. Does not implement Ward's (1963) clustering criterion.
\item "ward.D2" Implements that criterion (Murtagh and Legendre 2014). The dissimilarities are squared before cluster updating.
\item "single": Closely related to the minimal spanning tree. Adopts a ‘friends of friends’ clustering strategy.
\item "complete": Finds similar clusters.
\item "average": Same as UPGMA.
\item "mcquitty": Same as WPGMA.
\item "median": Same as WPGMC. Does not lead to a monotone distance measure, or equivalently the resulting dendrograms can have so called inversions or reversals which are hard to interpret, but note the trichotomies in Legendre and Legendre (2012).
\item "centroid": Same as UPGMC. Does not lead ... Legendre and Legendre (2012) (same as \code{median}).
}
Defaults to "ward.D".}
}
\value{
A lsit containing the filtered data frame, methods used, the plot selected, and other metadata.
}
\description{
This function plots correlational heatmaps based on several correlation and clustering methods. The plots are filtered based on the top samples/features/metabolites having the highest variability among groups.
}
\examples{
\dontrun{
plot_CorrelationHeatmap(data = results_from_perform_PreprocessingPeakData_function)
}

}
