% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Volcano.R
\name{plot_Volcano}
\alias{plot_Volcano}
\title{Plot Volcano Plot}
\usage{
plot_Volcano(
  PPData,
  FCData,
  CAData,
  arrangeLevels = NULL,
  fcUP = 2,
  fcDown = 0.5,
  adjpvalue = 0.05
)
}
\arguments{
\item{PPData}{List. This list must be a result from the \code{perform_PreprocessingPeakData} function.}

\item{FCData}{List. This list must be a result from the \code{perform_FoldChange} function.}

\item{CAData}{List. This list must be a result from the \code{perform_ComparativeAnalysis} function.}

\item{arrangeLevels}{Vector. Determines how the groups will be arranged. The format could be "c('group1', 'group2', ...)". Defaults to \code{NULL} which sorts the groups in alphabetical order. Suggests to be inputted (control, case1, case2) where case2 is worse than case1 e.g., Severe dengue than Non-severe dengue.}

\item{fcUP}{Numeric. The threshold for the upper fold change to be considered.}

\item{fcDown}{Numeric. The threshold for the lower fold change to be considered.}

\item{adjpvalue}{Numeric. The adjusted p-value threshold.}
}
\value{
Returns a list of data used in the analysis, volcano plot data, and the plot.
}
\description{
This function plots a volcano plot.
}
\examples{
\dontrun{
plot_Volcano(
  PPData    = results_from_perform_PreprocessingPeakData_function,
  FCData    = results_from_perform_FoldChange_function,
  CAData    = results_from_perform_ComparativeAnalysis_function
)
}

}
