% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{select_file_interactively}
\alias{select_file_interactively}
\title{Select File Interactively}
\usage{
select_file_interactively()
}
\description{
Select File Interactively
}
\keyword{internal}
