% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{transpose_back_to_original_format}
\alias{transpose_back_to_original_format}
\title{Transpose Back to Original Format}
\usage{
transpose_back_to_original_format(processed_data, first_col_name)
}
\description{
Transpose Back to Original Format
}
\keyword{internal}
