% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{transpose_for_processing}
\alias{transpose_for_processing}
\title{Transpose Data for Processing}
\usage{
transpose_for_processing(raw_data)
}
\description{
Transpose Data for Processing
}
\keyword{internal}
