% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{validate_and_locate_metadata}
\alias{validate_and_locate_metadata}
\title{Validate and Locate Metadata Rows}
\usage{
validate_and_locate_metadata(raw_data)
}
\description{
Validate and Locate Metadata Rows
}
\keyword{internal}
