% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate Data Structure}
\usage{
validate_data_structure(raw_data, allow_missing_optional)
}
\description{
Validate Data Structure
}
\keyword{internal}
