% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate Function Parameters}
\usage{
validate_parameters(merge_method, file_location)
}
\description{
Validate Function Parameters
}
\keyword{internal}
