% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{finalize_data_structure}
\alias{finalize_data_structure}
\title{Finalize Data Structure}
\usage{
finalize_data_structure(raw_data)
}
\description{
Finalize Data Structure
}
\keyword{internal}
