% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_PCA.R
\name{get_ScreeData}
\alias{get_ScreeData}
\title{Extract Scree Plot Data}
\usage{
get_ScreeData(multi_plots_obj, what = "variance", max_pc = NULL)
}
\arguments{
\item{multi_plots_obj}{Object returned from perform_PCA()}

\item{what}{Character. "variance" or "eigenvalue" to specify what values to return}

\item{max_pc}{Integer. Maximum number of PCs to include. If NULL, includes all.}
}
\value{
Data frame with PC information
}
\description{
Extract Scree Plot Data
}
