% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{initialize_results_object}
\alias{initialize_results_object}
\title{Initialize Results Object}
\usage{
initialize_results_object(
  file_location,
  sheet_name,
  validate_qc,
  allow_missing_optional,
  clean_names
)
}
\description{
Initialize Results Object
}
\keyword{internal}
