% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DataQualityCheck.R
\name{merge_duplicate_samples}
\alias{merge_duplicate_samples}
\title{Merge Duplicate Samples}
\usage{
merge_duplicate_samples(raw_data, method, verbose)
}
\description{
Merge Duplicate Samples
}
\keyword{internal}
