% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_ComparativeAnalysis.R
\name{perform_ComparativeAnalysis}
\alias{perform_ComparativeAnalysis}
\title{Perform Comparative Statistical Analysis on a Preprocessed Data}
\usage{
perform_ComparativeAnalysis(
  data,
  adjust_p_method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")[5],
  sort_p = TRUE,
  paired = FALSE,
  plot_iden_met = NULL
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{perform_PreprocessingPeakData} function.}

\item{adjust_p_method}{String. The p-value correction method. Defaults to "BH". Read more at \code{?p.adjust}.
\itemize{
\item "holm": Holm (1979). Less conservative than "bonferroni".
\item "hochberg": Hochberg (1988). Less conservative than "bonferroni".
\item "hommel": Hommel (1988). Less conservative than "bonferroni".
\item "bonferroni": Bonferroni correction. The p-values are multiplied by the number of comparisons.
\item "BH": Benjamini & Hochberg (1995). Less conservative than "bonferroni".
\item "BY": Benjamini & Yekutieli (2001). Less conservative than "bonferroni".
\item "fdr": Same as "BH".
\item "none":
}
Note: The first four methods are designed to give strong control of the family-wise error rate. There seems no reason to use the unmodified Bonferroni correction because it is dominated by Holm's method, which is also valid under arbitrary assumptions. Hochberg's and Hommel's methods are valid when the hypothesis tests are independent or when they are non-negatively associated ( Sarkar, 1998; Sarkar and Chang, 1997). Hommel's method is more powerful than Hochberg's, but the difference is usually small and the Hochberg p-values are faster to compute. The "BH" (aka "fdr") and "BY" methods of Benjamini, Hochberg, and Yekutieli control the false discovery rate, the expected proportion of false discoveries amongst the rejected hypotheses. The false discovery rate is a less stringent condition than the family-wise error rate, so these methods are more powerful than the others.}

\item{sort_p}{Boolean. If \code{TRUE} (default), sorts the adjusted p-values in ascending order.}

\item{paired}{Boolean. If \code{TRUE}, this instructs the function that the data is paired, not independent. All p-values are adjusted using Bonferroni and Hochberg method.}

\item{plot_iden_met}{Boolean. Setting this to \code{TRUE} will output a plot of the identified metabolites using \code{ggstatsplot} packagge, which means that statistical tests and p-values are displayed in the plot.}
}
\value{
A list of data frames and plots, if requested.
}
\description{
This function performs comparative analysis based on the characteristics of the data. This means that this function
checks for the assumptions first before proceeding with the statistical analysis. The tests include t-tests and
Analysis of Variance (ANOVA), and its types such as paired samples t-test, and their nonparametric counterparts such as
Mann-Whitney U test, Kruskal-Wallis test, etc.
}
\examples{
\dontrun{
perform_ComparativeAnalysis(data = results_from_perform_PreprocessingPeakData_function)
}

}
