% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performComparativeAnalysis.R
\name{performComparativeAnalysis}
\alias{performComparativeAnalysis}
\title{Perform Statistical Analysis on a Preprocessed Data}
\usage{
performComparativeAnalysis(
  data,
  sort_p = TRUE,
  paired = FALSE,
  plot_iden_met = NULL
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{performPreprocessingPeakData} function.}

\item{sort_p}{Boolean. If \code{TRUE} (default), sorts the adjusted p-values in ascending order.}

\item{paired}{Boolean. If \code{TRUE}, this instructs the function that the data is paired, not independent. All p-values are adjusted using Bonferroni and Hochberg method.}

\item{plot_iden_met}{Boolean. Setting this to \code{TRUE} will output a plot of the identified metabolites using \code{ggstatsplot} packagge, which means that statistical tests and p-values are displayed in the plot.}
}
\value{
A list of data frames and plots, if requested.
}
\description{
This function performs comparative analysis based on the characteristics of the data. This means that this function
checks for the assumptions first before proceeding with the statistical analysis. The tests include t-tests and
Analysis of Variance (ANOVA), and its types such as paired samples t-test, and their nonparametric counterparts such as
Mann-Whitney U test, Kruskal-Wallis test, etc.
}
\examples{
\dontrun{
performComparativeAnalysis(data = results_from_performPreprocessingPeakData_function)
}

}
