% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_Export2Excel.R
\name{perform_Export2Excel}
\alias{perform_Export2Excel}
\title{Perform Export Data Frames to Excel}
\usage{
perform_Export2Excel(
  results,
  folder_name = "Results_Folder",
  file_name = "Results Exports"
)
}
\arguments{
\item{results}{List. A list of results from \code{perform_PreprocessingPeakData}, \code{perform_DimensionReduction}, or \code{perform_ComparativeAnalysis}, etc. There must only be one at a time. The list may or may not contain a data frame.}

\item{folder_name}{String. Creates a folder in the current working directory where the Excel file will be exported.}

\item{file_name}{String. The file name of the results.}
}
\value{
An Excel file.
}
\description{
This function exports the data frames from a list to Excel. The data frames are detected automatically. All data frames
in a list are combined in one (1) Excel file.
}
\examples{
\dontrun{
perform_Export2Excel(
  results     = results_from_perform_Regression_function,
  folder_name = "Regression Results Folder",
  file_name   = "Regression_Results_Filename_no_extension")
}

}
