% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_PreprocessingPeakData.R
\name{perform_PreprocessingPeakData}
\alias{perform_PreprocessingPeakData}
\title{Comprehensive Data Preprocessing Pipeline}
\usage{
perform_PreprocessingPeakData(
  raw_data,
  outliers = NULL,
  filterMissing = 20,
  filterMissing_by_group = TRUE,
  filterMissing_includeQC = FALSE,
  denMissing = 5,
  driftBatchCorrection = TRUE,
  spline_smooth_param = 0,
  spline_smooth_param_limit = c(-1.5, 1.5),
  log_scale = TRUE,
  min_QC = 5,
  removeUncorrectedFeatures = TRUE,
  dataNormalize = "Normalization",
  refSample = NULL,
  groupSample = NULL,
  reference_method = "mean",
  dataTransform = "vsn",
  dataScalePCA = "meanSD",
  dataScalePLS = "mean2SD",
  filterMaxRSD = 30,
  filterMaxRSD_by = "EQC",
  filterMaxVarSD = 10,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{List. Quality-checked data from the \code{perform_DataQualityCheck} function.}

\item{outliers}{Vector. Biological samples and/or QC samples considered as outliers. Example format: \code{c('Sample1', 'Sample2', 'QC1', 'QC2', ...)}. Defaults to \code{NULL}.}

\item{filterMissing}{Numeric. Minimum percentage of missing values across all groups required to remove a feature.}

\item{filterMissing_by_group}{Boolean. Determines whether \code{filterMissing} should assess group-specific missingness before feature removal.}

\item{filterMissing_includeQC}{Boolean. If \code{FALSE} (default), QC samples are excluded when implementing \code{filterMissing}.}

\item{denMissing}{Numeric. Denominator value used in the fraction \code{1/denMissing} to replace missing values.}

\item{driftBatchCorrection}{Boolean. If \code{TRUE} (default), perform QC-RSC algorithm for signal drift and batch effect correction.}

\item{spline_smooth_param}{Numeric. Spline smoothing parameter ranging from 0 to 1.}

\item{spline_smooth_param_limit}{Vector. A vector of format \code{c(min, max)} for spline parameter limits.}

\item{log_scale}{Boolean. If \code{TRUE} (default), performs signal correction fit on log-scaled data.}

\item{min_QC}{Numeric. Minimum number of QC samples required for signal correction per batch.}

\item{removeUncorrectedFeatures}{Boolean. If \code{TRUE} (default), removes features that were not corrected by QCRSC due to insufficient QC samples meeting the min_QC threshold.}

\item{dataNormalize}{String. Data normalization method. Options:
\itemize{
\item \code{"none"}: No normalization
\item \code{"Normalization"}: Using the values from "Normalization" row
\item \code{"sum"}: By sum
\item \code{"median"}: By median
\item \code{"PQN1"}: By median of reference spectrum
\item \code{"PQN2"}: By reference sample supplied in \code{refSample}
\item \code{"groupPQN"}: By group in \code{c("SQC", "EQC", "QC")}, if both (default) then all QCs are considered as QC
\item \code{"quantile"}: By quantile
}
Default: "Normalization" (if present, otherwise, "sum")}

\item{refSample}{String. Reference sample for \code{dataNormalize = "PQN2"}.}

\item{groupSample}{String. Used only if \code{dataNormalize = "groupPQN"}.}

\item{reference_method}{String. Method for computing reference from QC samples in \code{dataNormalize = "quantile"}. Options:
\itemize{
\item \code{"mean"}: Default
\item \code{"median"}
}}

\item{dataTransform}{String. Data transformation method applied after \code{dataNormalize}. Options:
\itemize{
\item \code{"none"}: No transformation
\item \code{"log2"}: log base 2
\item \code{"log10"}: log base 10
\item \code{"sqrt"}: Square-root
\item \code{"cbrt"}: Cube-root
\item \code{"vsn"}: Variance Stabilizing Normalization
}}

\item{dataScalePCA}{String. Data scaling for PCA analysis. Options:
\itemize{
\item \code{"none"}: No data scaling
\item \code{"mean"}: Scale by mean (average)
\item \code{"meanSD"}: Scale by mean divided by standard deviation (SD)
\item \code{"mean2SD"}: Pareto-scaling. Scale by mean divided by square-root of SD. Always use this for PLS-type analysis.
}}

\item{dataScalePLS}{String. Data scaling for PLS analysis. Same options as \code{dataScalePCA}.}

\item{filterMaxRSD}{Numeric. Threshold for RSD filtering.}

\item{filterMaxRSD_by}{String. Which QC samples to use for RSD filtering. Options:
\itemize{
\item \code{"SQC"}: Filter by sample QC
\item \code{"EQC"}: Filter by extract QC (default)
\item \code{"both"}: Filter by by both SQC and EQC (or QC altogether)
}}

\item{filterMaxVarSD}{Numeric. Remove \verb{nth percentile} of features with lowest variability.}

\item{verbose}{Logical. Whether to print detailed progress messages. Default TRUE.}
}
\value{
A list containing results from all preprocessing steps.
}
\description{
Performs a complete data preprocessing workflow to prepare raw data for
downstream analysis. This function applies preprocessing steps sequentially
in the order specified by the parameters to ensure optimal data quality
and analytical readiness.
}
\seealso{
\code{\link{perform_DataQualityCheck}} for the data quality check
}
\author{
John Lennon L. Calorio
}
