% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performRegression.R
\name{performRegression}
\alias{performRegression}
\title{Perform Least Absolute Shrinkage and Selection Operator (LASSO) and Elastic Net Regression}
\usage{
performRegression(
  data,
  method = "enet",
  train_percent = 80,
  ref = NULL,
  lambda = "1se",
  remember = NULL
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{performPreprocessingPeakData} function.}

\item{method}{String or vector.
\itemize{
\item "lasso": Analyses data using LASSO only.
\item "enet": Analyses data using Elastic Net only.
\item c("lasso", "enet"): Analyses data using both LASSO Elastic Net. Needs to be a vector since 'Ridge Regression' will be added.
}
Defaults to "enet".}

\item{train_percent}{Numeric. The percent to be used in the training data set. The remaining will be used in testing. Selection will be done in a random manner. For consistent results, suggests to use the \code{remember} parameter.}

\item{ref}{String. The reference value. Defaults to the 1st level/category of the dependent variable if \code{NULL}.}

\item{lambda}{String. The lambda value.
\itemize{
\item "1se"
\item "min"
}
Defaults to "1se" for fewer features, lower risk of over fitting.}

\item{remember}{Numeric. This value will be used in the \code{set.seed(remember)} function.}
}
\value{
Data frame of regression results containing coefficients and odds ratios.
}
\description{
This function performs either or both the LASSO and/or Elastic Net Regression. Both are regularization
techniques used to prevent overfitting in linear regression models by adding penalty terms to the loss function.
LASSO uses L1 regularization to shrink the coefficients of less important features towards zero, potentially eliminating them entirely.
Elastic Net uses a combination of L1 and L2 regularization to address multicollinearity and perform feature selection simultaneously.
}
\examples{
\dontrun{
# Perform the function
myregression <- performRegression(
  data     = results_from_performPreprocessingPeakData_function,
  remember = 1234
  )

# View the results
View(myregression[["LASSO_coefOddsRatio"]])
View(myregression[["ElasticNet_coefOddsRatio"]])
myregression[["LASSO_confusionMatrix"]]$overall[1]
myregression[["ElasticNet_confusionMatrix"]]$overall[1]
}

}
