% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BeforeAfter.R
\name{plot_BeforeAfter}
\alias{plot_BeforeAfter}
\title{Plot Random Data Before and After Preprocessing}
\usage{
plot_BeforeAfter(
  data,
  scaled = "OPLS-DA",
  group_by = "Sample",
  n_random_samples = 30,
  n_random_features = 30
)
}
\arguments{
\item{data}{List. This list must be a result from the \code{perform_PreprocessingPeakData} function.}

\item{scaled}{String. Choose one below to plot one of the final results from the data preprocessing.
\itemize{
\item 'PCA: Plots results from PCA.
\item 'OPLS-DA': Plots results from OPLS-DA.
}
Defaults to "OPLS-DA".}

\item{group_by}{String. Whether to plot the "Samples" or the "Features".
\itemize{
\item 'Sample: Plots the samples.
\item 'Feature': Plots the features.
}
Defaults to "Sample".}

\item{n_random_samples}{Numeric. The number of random 'samples' to be created with box plots. If \code{NULL}, plots all. Defaults to 30, or the samples exceeds 30.}

\item{n_random_features}{Numeric. The number of random 'features' to be created with box plots. If \code{NULL}, plots all. Defaults to 30, or the features exceeds 30.}
}
\value{
A list of plots of the before and after data preprocessing.
}
\description{
This function will plot random samples or features the before and after data preprocessing.
}
\examples{
\dontrun{
plot_BeforeAfter(
  data     = results_from_perform_PreprocessingPeakData_function,
  scaled   = "OPLS-DA",
  group_by = "Sample"
)
}
}
